+++ collated/main/low5545.md
@@ -290,7 +290,7 @@ public class ExportCommandParser implements Parser<ExportCommand> {
         return new ImportXmlCommand(path);
     }
 ```
-###### \java\seedu\address\logic\parser\ParserUtil.java
+###### \java\seedu\address\logic\parser\util\ParserUtil.java
 ``` java
     /**
      * Parses and trims {@code filePath} if it is not empty.
