+++ src/main/java/seedu/address/logic/parser/event/EditEventParser.java
@@ -13,7 +13,7 @@ import seedu.address.logic.commands.event.EditEventCommand.EditEventDescriptor;
 import seedu.address.logic.parser.ArgumentMultimap;
 import seedu.address.logic.parser.ArgumentTokenizer;
 import seedu.address.logic.parser.Parser;
-import seedu.address.logic.parser.ParserUtil;
+import seedu.address.logic.parser.util.ParserUtil;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.property.exceptions.PropertyNotFoundException;
