+++ src/test/java/seedu/address/testutil/EditEventDescriptorBuilder.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 
 import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.logic.commands.event.EditEventCommand.EditEventDescriptor;
-import seedu.address.logic.parser.ParserUtil;
+import seedu.address.logic.parser.util.ParserUtil;
 import seedu.address.model.event.ReadOnlyEvent;
 import seedu.address.model.property.exceptions.PropertyNotFoundException;
